backend/requirements.txt
# Bird-App 2.0 Backend (Chirper schema)
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
sqlalchemy[asyncio]>=2.0.0
aiosqlite>=0.20.0
aiomysql>=0.2.0
asyncpg>=0.29.0
psycopg2-binary>=2.9.0
pydantic>=2.0
pydantic-settings>=2.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.9
httpx>=0.27.0
pytest>=8.0.0
pytest-asyncio>=0.24.0
pytest-cov>=6.0.0
python-dotenv>=1.0.0
